uber UX animal eaten management

+++ DominatingSpeciesSampleData/src/main/java/cz/muni/fi/pa165/dominatingspecies/sampledata/SampleDataLoadingFacadeImpl.java
@@ -18,16 +18,16 @@ import org.springframework.transaction.annotation.Transactional;
 @Named
 @Transactional
 public class SampleDataLoadingFacadeImpl implements SampleDataLoadingFacade {
-    
+
     @Inject
     private EnvironmentService environmentService;
-    
+
     @Inject
     private AnimalService animalService;
-    
+
     @Inject
     private AnimalEnvironmentService aeService;
-    
+
     @Inject
     private AnimalEatenService animalEatenService;
 
@@ -38,13 +38,13 @@ public class SampleDataLoadingFacadeImpl implements SampleDataLoadingFacade {
         Animal hamster = createAnimal("Škrečok", "Hlodavce", 45d, 6d);
         Animal rabbit = createAnimal("Zajac obyčajný", "Hlodavce", 21d, 3d);
         Animal camel = createAnimal("Ťava dvojhrbá", "Prežúvavce", 4d, 2d);
-        
+
         Environment home = createEnvironment("Domácnosti", "Domy, statky", 10l);
         Environment field = createEnvironment("Polia", "Obhospodarované polia", 100l);
         Environment forest = createEnvironment("Lesy", "Ihlicnate lesy", 40l);
         Environment desert = createEnvironment("Púšte", "Kamenisté púšte", 23l);
         Environment mountains = createEnvironment("Hory", "Vysoké hory", 190l);
-        
+
         AnimalEnvironment catHome = createAnimalEnvironment(cat, home, 1);
         AnimalEnvironment hamsterHome = createAnimalEnvironment(hamster, home, 1);
         AnimalEnvironment rabbitHome = createAnimalEnvironment(rabbit, home, 1);
@@ -54,52 +54,52 @@ public class SampleDataLoadingFacadeImpl implements SampleDataLoadingFacade {
         AnimalEnvironment mouseForest = createAnimalEnvironment(mouse, forest, 0.2);
         AnimalEnvironment mouseDesert = createAnimalEnvironment(mouse, desert, 0.9);
         AnimalEnvironment mouseMountains = createAnimalEnvironment(mouse, mountains, 0.8);
-        
-        AnimalEaten catMouse = createAnimalEaten(cat, mouse, 10l);
+
+        AnimalEaten catMouse = createAnimalEaten(cat, mouse, 10.0);
     }
-    
+
     private Animal createAnimal(String name, String species, double reproductionRate, double foodNeeded) {
         Animal animal = new Animal();
         animal.setName(name);
         animal.setSpecies(species);
         animal.setReproductionRate(reproductionRate);
         animal.setFoodNeeded(foodNeeded);
-        
+
         animalService.create(animal);
-        
+
         return animal;
     }
-    
+
     private Environment createEnvironment(String name, String description, long maxAnimalCount) {
         Environment environment = new Environment();
         environment.setName(name);
         environment.setDescription(description);
         environment.setMaxAnimalCount(maxAnimalCount);
-        
+
         environmentService.create(environment);
-        
+
         return environment;
     }
-    
+
     private AnimalEnvironment createAnimalEnvironment(Animal animal, Environment environment, double percentage) {
         AnimalEnvironment animalEnvironment = new AnimalEnvironment();
         animalEnvironment.setAnimal(animal);
         animalEnvironment.setEnvironment(environment);
         animalEnvironment.setPercentage(percentage);
-        
+
         aeService.create(animalEnvironment);
-        
+
         return animalEnvironment;
     }
-    
-    private AnimalEaten createAnimalEaten(Animal predator, Animal prey, long animalCount) {
+
+    private AnimalEaten createAnimalEaten(Animal predator, Animal prey, Double animalCount) {
         AnimalEaten eaten = new AnimalEaten();
         eaten.setPredator(predator);
         eaten.setPrey(prey);
         eaten.setAnimalCount(animalCount);
-        
+
         animalEatenService.createAnimalEaten(eaten);
-        
+
         return eaten;
     }
 }
